Address review: A single touchscreen now automatically picked up; long press times validated to be <= 65535ms ifdef/endifs corrected.

+++ esphome/components/lvgl/lvgl_esphome.h
@@ -28,7 +28,7 @@
 #endif
 #ifdef USE_LVGL_TOUCHSCREEN
 #include "esphome/components/touchscreen/touchscreen.h"
-#endif
+#endif  // USE_LVGL_TOUCHSCREEN
 
 namespace esphome {
 namespace lvgl {
@@ -36,14 +36,14 @@ namespace lvgl {
 extern lv_event_code_t lv_custom_event;  // NOLINT
 #ifdef USE_LVGL_COLOR
 inline lv_color_t lv_color_from(Color color) { return lv_color_make(color.red, color.green, color.blue); }
-#endif
+#endif  // USE_LVGL_COLOR
 #if LV_COLOR_DEPTH == 16
 static const display::ColorBitness LV_BITNESS = display::ColorBitness::COLOR_BITNESS_565;
 #elif LV_COLOR_DEPTH == 32
 static const display::ColorBitness LV_BITNESS = display::ColorBitness::COLOR_BITNESS_888;
-#else
+#else   // LV_COLOR_DEPTH
 static const display::ColorBitness LV_BITNESS = display::ColorBitness::COLOR_BITNESS_332;
-#endif
+#endif  // LV_COLOR_DEPTH
 
 // Parent class for things that wrap an LVGL object
 class LvCompound final {
@@ -147,7 +147,7 @@ class LvglComponent : public PollingComponent {
 #ifdef USE_LVGL_TOUCHSCREEN
 class LVTouchListener : public touchscreen::TouchListener, public Parented<LvglComponent> {
  public:
-  LVTouchListener(uint32_t long_press_time, uint32_t long_press_repeat_time) {
+  LVTouchListener(uint16_t long_press_time, uint16_t long_press_repeat_time) {
     lv_indev_drv_init(&this->drv_);
     this->drv_.long_press_repeat_time = long_press_repeat_time;
     this->drv_.long_press_time = long_press_time;
@@ -177,7 +177,7 @@ class LVTouchListener : public touchscreen::TouchListener, public Parented<LvglC
   touchscreen::TouchPoint touch_point_{};
   bool touch_pressed_{};
 };
-#endif
+#endif  // USE_LVGL_TOUCHSCREEN
 
 #ifdef USE_LVGL_KEY_LISTENER
 class LVEncoderListener : public Parented<LvglComponent> {
